Reorganised client folder.
