+++ client/tx/txRoutes.go
@@ -1,61 +1,24 @@
-package client
+package tx
 
 import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"github.com/cosmos/cosmos-sdk/client/context"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/rest"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/client/utils"
+	genutilrest "github.com/cosmos/cosmos-sdk/x/genutil/client/rest"
+	"github.com/gorilla/mux"
+	utils2 "github.com/ixofoundation/ixo-blockchain/client/utils"
+	"github.com/ixofoundation/ixo-blockchain/x/ixo"
 	"github.com/ixofoundation/ixo-blockchain/x/project"
 	"io/ioutil"
 	"net/http"
 	"strconv"
 	"strings"
-	"time"
-
-	"github.com/cosmos/cosmos-sdk/client/context"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/rest"
-	"github.com/gorilla/mux"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	core "github.com/tendermint/tendermint/rpc/core/types"
-
-	genutilrest "github.com/cosmos/cosmos-sdk/x/genutil/client/rest"
-	"github.com/ixofoundation/ixo-blockchain/x/ixo"
 )
-
-func QueryTxCmd(cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-
-		Use:   "tx [hash]",
-		Short: "Query for a transaction by hash in a committed block",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-
-			output, err := QueryTx(cliCtx, args[0])
-			if err != nil {
-				return err
-			}
-
-			if output.Empty() {
-				return fmt.Errorf("No transaction found with hash %s", args[0])
-			}
-
-			return cliCtx.PrintOutput(output)
-		},
-	}
-
-	cmd.Flags().StringP(flags.FlagNode, "n", "tcp://localhost:26657", "Node to connect to")
-	_ = viper.BindPFlag(flags.FlagNode, cmd.Flags().Lookup(flags.FlagNode))
-	cmd.Flags().Bool(flags.FlagTrustNode, false, "Trust connected full node (don't verify proofs for responses)")
-	_ = viper.BindPFlag(flags.FlagTrustNode, cmd.Flags().Lookup(flags.FlagTrustNode))
-
-	return cmd
-}
 
 func RegisterTxRoutes(cliCtx context.CLIContext, r *mux.Router) {
 	r.HandleFunc("/txs/{hash}", QueryTxRequestHandlerFn(cliCtx)).Methods("GET")
@@ -76,7 +39,7 @@ func QueryTxRequestHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 			return
 		}
 
-		output, err := QueryTx(cliCtx, hashHexStr)
+		output, err := utils2.QueryTx(cliCtx, hashHexStr)
 		if err != nil {
 			if strings.Contains(err.Error(), hashHexStr) {
 				rest.WriteErrorResponse(w, http.StatusNotFound, err.Error())
@@ -147,41 +110,6 @@ func QueryTxsRequestHandlerFn(cliCtx context.CLIContext) http.HandlerFunc {
 
 		rest.PostProcessResponseBare(w, cliCtx, searchResult)
 	}
-}
-
-func QueryTx(cliCtx context.CLIContext, hashHexStr string) (sdk.TxResponse, error) {
-	hash, err := hex.DecodeString(hashHexStr)
-	if err != nil {
-		return sdk.TxResponse{}, err
-	}
-
-	node, err := cliCtx.GetNode()
-	if err != nil {
-		return sdk.TxResponse{}, err
-	}
-
-	resTx, err := node.Tx(hash, !cliCtx.TrustNode)
-	if err != nil {
-		return sdk.TxResponse{}, err
-	}
-
-	if !cliCtx.TrustNode {
-		if err = ValidateTxResult(cliCtx, resTx); err != nil {
-			return sdk.TxResponse{}, err
-		}
-	}
-
-	resBlocks, err := getBlocksForTxResults(cliCtx, []*core.ResultTx{resTx})
-	if err != nil {
-		return sdk.TxResponse{}, err
-	}
-
-	out, err := formatTxResult(cliCtx.Codec, resTx, resBlocks[resTx.Height])
-	if err != nil {
-		return out, err
-	}
-
-	return out, nil
 }
 
 type BroadcastReq struct {
@@ -307,54 +235,4 @@ func SignDataRequest(cliCtx context.CLIContext) http.HandlerFunc {
 
 		rest.PostProcessResponseBare(w, cliCtx, output)
 	}
-}
-
-func getBlocksForTxResults(cliCtx context.CLIContext, resTxs []*core.ResultTx) (map[int64]*core.ResultBlock, error) {
-	node, err := cliCtx.GetNode()
-	if err != nil {
-		return nil, err
-	}
-
-	resBlocks := make(map[int64]*core.ResultBlock)
-	for _, resTx := range resTxs {
-		if _, ok := resBlocks[resTx.Height]; !ok {
-			resBlock, err := node.Block(&resTx.Height)
-			if err != nil {
-				return nil, err
-			}
-
-			resBlocks[resTx.Height] = resBlock
-		}
-	}
-
-	return resBlocks, nil
-}
-
-func formatTxResult(cdc *codec.Codec, resTx *core.ResultTx, resBlock *core.ResultBlock) (sdk.TxResponse, error) {
-	tx, err := parseTx(cdc, resTx.Tx)
-	if err != nil {
-		return sdk.TxResponse{}, err
-	}
-
-	return sdk.NewResponseResultTx(resTx, tx, resBlock.Block.Time.Format(time.RFC3339)), nil
-}
-
-func parseTx(cdc *codec.Codec, txBytes []byte) (sdk.Tx, error) {
-	return ixo.DefaultTxDecoder(cdc)(txBytes)
-}
-
-func ValidateTxResult(cliCtx context.CLIContext, resTx *core.ResultTx) error {
-	if !cliCtx.TrustNode {
-		check, err := cliCtx.Verify(resTx.Height)
-		if err != nil {
-			return err
-		}
-
-		err = resTx.Proof.Validate(check.Header.DataHash)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
